requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==1.0.15
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
cryptography==45.0.5
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
durationpy==0.9
emoji==2.14.1
executing==2.1.0
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.10.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.66.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
json5==0.9.28
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kubernetes==31.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.6
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.47.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.3
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pdfminer.six==20250506
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
posthog==3.7.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.3
pypdf==5.8.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-json-logger==2.0.7
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
Send2Trash==1.8.3
setuptools==75.5.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.41.3
streamlit==1.47.0
sympy==1.13.3
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.13.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
unstructured==0.18.9
unstructured-client==0.39.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
yarl==1.20.1
zipp==3.21.0
zstandard==0.23.0
